refactor(Types): ♻️ Redistributed global interfaces into domain-specific modules.     -   Moved navigation interfaces to `navigation.d.ts`, specifically     moving `Anchor`, `NavProps`, and `PaginationProps` to centralize     navigation-related types and enhance modularity.
    -   Moved media interfaces to `media.d.ts`, moving `IconProps`,
    `FigureItem`, and `FigureProps` to create a dedicated space for
    media-related types.

    -   Deconstructed the monolithic `props.d.ts`, which improves
    project organization and makes it easier for developers to locate
    specific type definitions.

    -   Reduced cross-module clutter, which minimizes circular
    dependencies and ensures that imports are more intuitive and
    contextually relevant.

    -   Enhanced type discovery, providing a cleaner architectural
    foundation for future scaling of both navigation and media-heavy
    features.
